Split DB's struct from Tauri plugin
