+++ src-tauri/src/libs/mod.rs
@@ -1,3 +1,4 @@
+pub mod database;
 pub mod error;
 pub mod events;
 pub mod file_associations;
